requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.20
Django==1.10.5
djangorestframework==3.5.3
kombu==3.0.37
psycopg2==2.5.3
pytz==2016.10
redis==2.10.5
djangorestframework-jwt==1.8.0
dj-database-url==0.4.1
gunicorn==19.6.0
PyJWT==1.4.1
requests==2.10.0
django-nomad-base-accounts==2.3.10
django-extensions==1.6.7
django-crispy-forms==1.6.0
django-storages==1.5.0
